requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Marshmallow==1.2.1
marshmallow==3.22.0
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
reportlab==4.2.5
